NF(TST): externals_use_cassette to make custom external remotes use cassettes
