+++ datalad/tests/utils.py
@@ -52,41 +52,6 @@ from . import _TEMP_PATHS_GENERATED
 # temp paths used by clones
 _TEMP_PATHS_CLONES = set()
 
-try:
-    # TEMP: Just to overcome problem with testing on jessie with older requests
-    # https://github.com/kevin1024/vcrpy/issues/215
-    import vcr.patch as _vcrp
-    import requests as _
-    try:
-        from requests.packages.urllib3.connectionpool import HTTPConnection as _a, VerifiedHTTPSConnection as _b
-    except ImportError:
-        def returnnothing(*args, **kwargs):
-            return()
-        _vcrp.CassettePatcherBuilder._requests = returnnothing
-
-    from vcr import use_cassette as _use_cassette, VCR as _VCR
-
-    def use_cassette(path, return_body=None, **kwargs):
-        """Adapter so we could create/use custom use_cassette with custom parameters
-        """
-        if return_body is not None:
-            my_vcr = _VCR(before_record_response=lambda r: dict(r, body={'string': return_body.encode()}))
-            return my_vcr.use_cassette(path, **kwargs)  # with a custom response
-        else:
-            return _use_cassette(path, **kwargs)  # just a straight one
-
-except Exception as exc:
-    if not isinstance(exc, ImportError):
-        # something else went hairy (e.g. vcr failed to import boto due to some syntax error)
-        lgr.warning("Failed to import vcr, no cassettes will be available: %s", exc_str(exc, limit=10))
-    # If there is no vcr.py -- provide a do nothing decorator for use_cassette
-    def use_cassette(*args, **kwargs):
-        def do_nothing_decorator(t):
-            @wraps(t)
-            def wrapper(*args, **kwargs):
-                return t(*args, **kwargs)
-            return wrapper
-        return do_nothing_decorator
 
 def skip_if_no_module(module):
     try:
